pyproject.toml
[project]
name = "directory-ocr"
version = "0.1.0"
description = "Automated file processing pipeline with OCR and LLM-based text extraction"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "watchdog>=4.0.0",
    "pypdf2>=3.0.0",
    "requests>=2.31.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
directory-ocr = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = []
